Add memory regions to debug_lockstep script to prevent Qemu crash
GDB tries to be helpful, and does things like dereference the lr
register. But sometimes the lr gets pointed at an invalid location,
and reading from it causes Qemu to report a "hardware" error due
to a bad read. To prevent GDB from reading stuff, mark the entire
IO region as "write only".

This should be all we need, but in the future, if we need to read
directly from the IO region, we can disable and reenable the
region attributes.