Add debug waves to the datapath
To facilitate lockstep debugging based on the program counter, as
well as interrupt and memory forwarding, add waves to datapath.sv
that track whether instructions are valid (not flushed) and when
they are running micro-ops. Also add waves that track when the
execute and writeback stages "advance", i.e. when a valid
instruction in the correspoinding stage is complete.