Add IO forwarding shim to bus
In order to lockstep test with Qemu in the presence of
non-deterministic elements such as IO, interrupts, etc., we need
to forward Qemu's state into ModelSim. This has two main parts:
 - Forwarding IO accesses so that ModelSim and Qemu see same data
 - Forwarding IRQ and FIQ lines so that they trigger interrupts
   together
To accomplish this, add io_fwd_shim, a module that uses a queue to
return the appropriate memory accesses, and has functions that
enable adding reads to the queue and setting the IRQ and FIQ lines.

Also, to make this shim work with all IO, change the ahb setup to
route anything above 0x10000000 to the shim, and disable the
implementations of the pic and sic (which will be handled through
Qemu when we are lockstepping).