Add Python wrapper for LEG interaction, and update lockstep.tcl
This commit provides a Python-friendly wrapper (leg.py) that
handles all communication with ModelSim, including both lockstep
commands and checkpointing. It also allows a GUI mode, if we
want to open ModelSim in the GUI for debugging. Accordingly, some
changes were made to lockstep.tcl to extend it for general use.

Eventually, leg.py and lockstep.tcl will be the only pathway for
communication between Python and ModelSim.