Make new lockstepping the default, and integrate with GDB.
This commit integrates the new debugging functionality into the
existing GDB setup, and adds new options such as the GUI mode. Note
that the new debug interface is NOT compatible with the old one, so
the format of bugs and such are different.

Also delete files that are no longer necessary