Add option for spontaneous interrupts (when not implemented)
The original implementation of the lockstepping assumes that
interrupts are implemented, and is structured in a way such that
they happen in ModelSim first. However, when interrupts are not
implemented, we cannot test other functionality that depends on
them. This adds an option so that qemu will trigger interrupts as
normal, and then mark the resulting bug as a special "Interrupts
Not Implemented" bug, without waiting for ModelSim before
triggering the interrupt.