Add interrupt automation to unit tests
This allows us to debug interrupts in a semi-reliable fashion. In
a test file, if you use labels `interrupt_0:`, `interrupt_1:`, etc.
and include a .elf file with your .bin file, the lockstep debugger
will automatically trigger the IRQ line to go high. Note that any
test that uses this must first enable interrupts on both the
processor and the UART.