Use function-based memory loading mechanism
Original implementation of dmem.sv was subject to conflicts when
two different versions of the debug script started at the same
time, since they would both try to overwrite and the use
sim/simtest.dat. Replace it with a function called from tcl that
specifies the file to load after compile time. Also
 - Add environment variable to enable automatic loading of the
   original sim/simtest.dat file
 - Configure debug_lockstep to automatically clear memory when
   running with Linux, but not when running a unit test.