Modify compilation to happen before simulation
When running debug scripts in parallel, compilation was becoming
a problem between multiple instances, since they were all using
the same working directory. The temporary solution was to make
a new working directory for each script, but that took a long
time and was slower. By compiling before we run any of the
simulation and removing the compilation step from the tcl files,
this resolves the issue without having to use multiple working
directories.