Put limit on number of frames displayed by GDB
In particular functions in the kernel, stack frames seem to be
recursive. One example is __irq_svc. When you run "where" while at
one of these frames, the command never terminates, and just repeats
the same location thousands of times. Put a hard cap (20) on the
amount displayed by GDB.